+++ src/Settings.ts
@@ -1,6 +1,6 @@
 import * as vscode from 'vscode';
+import { ListsSortTypes, ObjectsSortTypes, SortModes, ValueTypeOrder } from './enum';
 
-export type SortType = 'Key' | 'Key Length' | 'Value' | 'Value Length' | 'Value Type';
 export class Settings {
   static get configuration() {
     return vscode.workspace.getConfiguration('sort-json.settings');
@@ -11,49 +11,53 @@ export class Settings {
   static setSettings(key: string, val: any, isGlobal = true) {
     return Settings.configuration.update(key, val, isGlobal);
   }
-  static get contextMenu() {
-    return (Settings.getSettings('contextMenu') as object) || {};
+
+  static get sortMode() {
+    return Settings.getSettings('sortMode') as SortModes || SortModes.both;
   }
-  static get orderOverride() {
-    return (Settings.getSettings('orderOverride') as string[]) || [];
+
+  static get sortLevel() {
+    return Settings.getSettings('sortLevel') as number || -1;
   }
+  static set sortLevel(level: number) {
+    Settings.setSettings('sortLevel', level);
+  }
+
+  static get objectSortType() {
+    return Settings.getSettings('objectSortType') as ObjectsSortTypes || ObjectsSortTypes.key;
+  }
+  static set objectSortType(value: ObjectsSortTypes) {
+    Settings.setSettings('objectSortType', value);
+  }
+
+  static get listSortType() {
+    return Settings.getSettings('listSortType') as ListsSortTypes || ListsSortTypes.value;
+  }
+  static set listSortType(value: ListsSortTypes) {
+    Settings.setSettings('listSortType', value);
+  }
+
+  static get sortValueTypeOrder() {
+    return Settings.getSettings('sortValueTypeOrder') as ValueTypeOrder[];
+  }
+
   static get isCaseSensitive() {
     return Settings.getSettings('isCaseSensitive') as boolean;
   }
   static set isCaseSensitive(value: boolean) {
     Settings.setSettings('isCaseSensitive', value);
   }
-  static get sortType() {
-    return Settings.getSettings('sortType') as SortType;
-  }
-  static set sortType(value: SortType) {
-    Settings.setSettings('sortType', value);
-  }
-  static get sortValueTypeOrder() {
-    return (
-      (Settings.getSettings('sortValueTypeOrder') as string[]) || [
-        'boolean',
-        'null',
-        'number',
-        'string',
-        'array',
-        'object',
-      ]
-    );
-  }
-  static get deepSortLevel() {
-    return Settings.getSettings('deepSortLevel') as number;
-  }
-  static get ignoreArraysOnDeepSort() {
-    return Settings.getSettings('ignoreArraysOnDeepSort') as boolean;
-  }
-  static get ignoreObjectsOnDeepSort() {
-    return Settings.getSettings('ignoreObjectsOnDeepSort') as boolean;
+
+  static get orderOverrideKeys() {
+    return (Settings.getSettings('orderOverrideKeys') as string[]) || [];
   }
   static get excludePaths() {
     return (Settings.getSettings('excludePaths') as string[]) || [];
   }
-  static get customComparisons() {
-    return (Settings.getSettings('customComparisons')) as Array<{ description: string, comparison: string }>;
+  static get customSortComparisons() {
+    return (Settings.getSettings('customSortComparisons')) as Array<{ description: string, comparison: string }>;
+  }
+  static get contextMenu() {
+    return (Settings.getSettings('contextMenu') as object) || {};
   }
 }