+++ src/extension.ts
@@ -3,30 +3,32 @@
 import * as vscode from 'vscode';
 import { Commands } from './enum';
 import { Settings } from './Settings';
-import { setCaseSensitive, setSortType, sortJSON, sortJSONByCustomComparison } from './sortJSON';
+import SortJSON from './sortJSON';
 
 // this method is called when your extension is activated
 // your extension is activated the very first time the command is executed
 export function activate(context: vscode.ExtensionContext) {
-  // Sort JSON
-  context.subscriptions.push(vscode.commands.registerCommand(Commands.SORT_JSON, () => sortJSON(false, false)));
-  // Sort JSON Deep
-  context.subscriptions.push(vscode.commands.registerCommand(Commands.SORT_JSON_DEEP, () => sortJSON(true, false)));
-  // Sort JSON Reverse
-  context.subscriptions.push(vscode.commands.registerCommand(Commands.SORT_JSON_REVERSE, () => sortJSON(false, true)));
-  // Sort JSON Deep Reverse
-  context.subscriptions.push(vscode.commands.registerCommand(Commands.SORT_JSON_DEEP_REVERSE, () => sortJSON(true, true)));
-  // Sort JSON By Custom Comparison
-  context.subscriptions.push(vscode.commands.registerCommand(Commands.SORT_JSON_BY_CUSTOM_COMPARISON, sortJSONByCustomComparison));
-  
-  // Set Sort Type
-  context.subscriptions.push(vscode.commands.registerCommand(Commands.SET_SORT_TYPE, () => setSortType()));
+
+  const sortJSON = new SortJSON();
+
+  // Sort Ascending
+  context.subscriptions.push(vscode.commands.registerCommand(Commands.ascendingSort, () => sortJSON.sort()));
+  // Sort Descending
+  context.subscriptions.push(vscode.commands.registerCommand(Commands.descendingSort, () => sortJSON.sort(true)));
+  // Do Custom Sort
+  context.subscriptions.push(vscode.commands.registerCommand(Commands.customSort, () => sortJSON.sort(false, true)));
+  // Set Sort Level
+  context.subscriptions.push(vscode.commands.registerCommand(Commands.setSortLevel, () => sortJSON.setSortLevel()));
+  // Set Object Sort Type
+  context.subscriptions.push(vscode.commands.registerCommand(Commands.setObjectSortType, () => sortJSON.setObjectSortType()));
+  // Set List Sort Type
+  context.subscriptions.push(vscode.commands.registerCommand(Commands.setListSortType, () => sortJSON.setListSortType()));
   // Toggle sort case sensitive
-  context.subscriptions.push(vscode.commands.registerCommand(Commands.TOGGLE_CASE_SENSITIVE, () => setCaseSensitive(!Settings.isCaseSensitive)));
+  context.subscriptions.push(vscode.commands.registerCommand(Commands.toggleCaseSensitive, () => sortJSON.setCaseSensitive(!Settings.isCaseSensitive)));
   // Set sort case sensitive
-  context.subscriptions.push(vscode.commands.registerCommand(Commands.SET_CASE_SENSITIVE, () => setCaseSensitive(true)));
+  context.subscriptions.push(vscode.commands.registerCommand(Commands.setCaseSensitive, () => sortJSON.setCaseSensitive(true)));
   // Set sort case insensitive
-  context.subscriptions.push(vscode.commands.registerCommand(Commands.SET_CASE_INSENSITIVE, () => setCaseSensitive(false)));
+  context.subscriptions.push(vscode.commands.registerCommand(Commands.setCaseInSensitive, () => sortJSON.setCaseSensitive(false)));
 }
 
 // this method is called when your extension is deactivated